Renamed to StaticProperty, moved to lib.py
